move App to src folder
