+++ src/App.js
@@ -1,10 +1,10 @@
 import React from "react";
 import { BrowserRouter as Router } from "react-router-dom";
-import PlanetProvider from "../store/PlanetProvider";
+import PlanetProvider from "./store/PlanetProvider";
 import "bootstrap/dist/js/bootstrap";
-import Navigation from "../layouts/Navigation";
-import Main from "../layouts/Main";
-import Footer from "../layouts/Footer";
+import Navigation from "./layouts/Navigation";
+import Main from "./layouts/Main";
+import Footer from "./layouts/Footer";
 
 const App = () => {
   return (
